[PATCH] introduce nonheritable mounts
These mounts are not copied in new mount namespaces.

These mounts can help to solve a problem like this:
An application can create a temporary mount or a mount tree for a short
perdiod in a temporary directory. Currently this application can't be
sure, that it will be able to remove the directory, because someone else
can create a new namespace, where mounts will be still mounted in this
directory.