kernel: add a netlink interface to get information about tasks
The proc file system is convenient for humans. It provides a lot of
information about processes in a human-readable format (text files).

When we want to get information from applications, we need to parse these
text files. The proc file system is slow, because we need to open a
file, read data from it and close the file. These operations are
expensive. If we want to get a few properties, they can be placed in
different files.

The task diag interfase is based on netlink sockets and looks like
socket-diag, which is used to get information about sockets.

If we want to request information about tasks, we need to send a request:

struct task_diag_pid {
	__u64	show_flags;	/* specify requred information*/
	__u32	num;		/* a number of tasks */
	__u32	pids[0];	/* an array of tasks */
};

A respone is a set of netlink messages. Each message describes one task. All
task properties are devided on groups. A message contains attributes for each
requested group, which are specifies in show_flags.  For example, if we request
the TASK_DIAG_SHOW_PIDS group, a response will contain the TASK_DIAG_PID
attribute, which described by the task_diag_pids structure.

struct task_diag_pids {
	__u32	tgid;
	__u32	ngid;
	__u32	pid;
	__u32	ppid;
	__u32	tpid;
};

The task diag is much faster than the proc file system. We don't need to
create a new file descriptor for each task. We need to send a request
and get a response. It allows to get information for a few task for one
request-response iteration.

I have compared performance of procfs and task-diag for the
"ps ax -o pid,ppid" command.

$ # A number of tasks
$ ps ax -o pid,ppid | wc -l
10348

$ time ps ax -o pid,ppid > /dev/null

real	0m1.073s
user	0m0.086s
sys	0m0.903s

$ time ./task_diag_all > /dev/null

real	0m0.037s
user	0m0.004s
sys	0m0.020s

$ perf stat -e syscalls:sys_exit* -- ps ax -o pid,ppid  2>&1 | grep syscalls | sort -n -r | head -n 5
            20,713      syscalls:sys_exit_open
            20,710      syscalls:sys_exit_close
            20,708      syscalls:sys_exit_read
            10,348      syscalls:sys_exit_newstat
                31      syscalls:sys_exit_write

$ perf stat -e syscalls:sys_exit* -- ./task_diag_all  2>&1 | grep syscalls | sort -n -r | head -n 5
               114      syscalls:sys_exit_recvfrom
                49      syscalls:sys_exit_write
                 8      syscalls:sys_exit_mmap
                 4      syscalls:sys_exit_mprotect
                 3      syscalls:sys_exit_newfstat

Signed-off-by: Andrey Vagin <avagin@openvz.org>