pid: restore the old behaviour of the ns_last_pid sysctl
CRIU uses ns_last_pid to fork a process with a specified pid. For
example, if we need to create a process with the pid of 10000,
we write 9999 into /proc/sys/kernel/ns_last_pid

$ echo 9999 > /proc/sys/kernel/ns_last_pid; sh -c 'echo $$'
10000

This behaviour has been broken and now if we write 9999 to ns_last_pid,
a process will get the pid 9999. This patch restores the old behaviour.

fixes: ("pid: replace pid bitmap implementation with IDR API")
Cc: Gargi Sharma <gs051095@gmail.com>
Cc: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Andrei Vagin <avagin@openvz.org>